feat: add cron job for low treasury balance
